fix(states): add missing state from ActivityState
That enumeration contains all possible values of the state field of an Activity.

Originaly it comes from the enum org.bonitasoft.engine.bpm.flownode.ActivityStates in bonita-engine.
However that enum is not complete, the 2 states "completing activity with boundary" and "aborting activity with boundary"  are missing

Relates to [BTT-17](https://bonitasoft.atlassian.net/browse/BTT-17)